devide test/elements.js on parts by methods
